Cleaned up app structure: removed, merged, renamed and reordered files, only entry points in root app dir now; JWT validation is now done inside 'rank_services()' to include entry points other than the API service
