+++ TEStribute/__init__.py
@@ -5,18 +5,12 @@ import logging
 import os
 from typing import (Iterable, Mapping, Optional, Union)
 
+from werkzeug.exceptions import Unauthorized
+
 from TEStribute import models
-from TEStribute import rank_order
-from TEStribute.access_drs import fetch_drs_objects_metadata
-from TEStribute.access_tes import fetch_tes_task_info
-from TEStribute.costs import estimate_costs
-from TEStribute.config.parse_config import config_parser
-from TEStribute.distances import estimate_distances
-from TEStribute.errors import (ResourceUnavailableError, ValidationError)
-from TEStribute.log.logging_functions import log_yaml
-from TEStribute.log import setup_logger
-from TEStribute.times import estimate_times
-from TEStribute.utils import get_valid_service_combinations
+from TEStribute.config import config_parser
+from TEStribute.errors import ValidationError
+from TEStribute.log import (log_yaml, setup_logger)
 
 # Set up logging
 log_file = os.path.abspath(
@@ -77,27 +71,6 @@ def rank_services(
                 }
             where [drs_id] entries are taken from parameter `drs_ids`.
     """
-    # Create Request object
-    log_yaml(
-        header="=== USER INPUT ===",
-        level=logging.DEBUG,
-        logger=logger,
-        drs_ids=drs_ids,
-        drs_uris=drs_uris,
-        mode=mode,
-        resource_requirements=resource_requirements,
-        tes_uris=tes_uris,
-    )
-    request = models.Request(
-        drs_ids=drs_ids,
-        drs_uris=drs_uris,
-        mode=mode,        
-        resource_requirements=models.ResourceRequirements(
-            **resource_requirements
-        ),
-        tes_uris=tes_uris,
-    )
-
     # Parse config file
     log_yaml(
         header="=== CONFIG ===",
@@ -111,6 +84,33 @@ def rank_services(
         config=config,
     )
 
+    # Create Request object
+    log_yaml(
+        header="=== USER INPUT ===",
+        level=logging.DEBUG,
+        logger=logger,
+        drs_ids=drs_ids,
+        drs_uris=drs_uris,
+        mode=mode,
+        resource_requirements=resource_requirements,
+        tes_uris=tes_uris,
+        jwt=jwt,
+    )
+    try:
+        request = models.Request(
+            drs_ids=drs_ids,
+            drs_uris=drs_uris,
+            mode=mode,        
+            resource_requirements=models.ResourceRequirements(
+                **resource_requirements
+            ),
+            tes_uris=tes_uris,
+            jwt=jwt,
+            jwt_config=config["security"]["jwt"],
+        )
+    except Unauthorized:
+        raise
+
     # Validate input parameters
     log_yaml(
         header="=== VALIDATION ===",
@@ -122,150 +122,134 @@ def rank_services(
     except ValidationError:
         raise
 
-    # Get metadata for input objects
-    # TODO: Use model/class for return object
-    # TODO: Add to constructor of Response
-    try:
-        drs_object_info = fetch_drs_objects_metadata(
-            drs_uris=request.drs_uris,
-            drs_ids=request.drs_ids,
-            timeout=config["timeout"]
-        )
-    except ResourceUnavailableError:
-        raise
-
-    # Get TES task info for resource requirements
-    # TODO: Use model/class for return object
-    # TODO: Add to constructor of Response
-    try:
-        tes_task_info = fetch_tes_task_info(
-            tes_uris=request.tes_uris,
-            resource_requirements=request.resource_requirements,
-            timeout=config["timeout"]
-        )
-    except ResourceUnavailableError:
-        raise
-
-    # Get valid TES-DRS combinations
-    # TODO: Use model/class for return object
-    # TODO: Add to constructor of Response
-    valid_service_combos = get_valid_service_combinations(
-        task_info=tes_task_info,
-        object_info=drs_object_info,
+    # Create Response object
+    response = models.Response(
+        request=request,
+        timeout=config["timeout"],
     )
 
-    # Compute distances
-    # TODO: Add to constructor of Response
-    tes_object_distances = estimate_distances(
-        combinations=valid_service_combos,
-    )
-
-    # Filter service combinations
-    # TODO: Add to constructor of Service Combinations
-
-    # Compute cost estimates
-    # TODO: Should be method of class ServiceCombinations
-    if models.Mode["cost"].value <= request.mode < models.Mode["time"].value:
-        tes_costs = estimate_costs(
-            task_info=tes_task_info,
-            object_info=drs_object_info,
-            distances=tes_object_distances,
-        )
-    else: tes_costs = {}
-
-    # Compute time estimates
-    # TODO: Should be method of class ServiceCombinations
-    if models.Mode["cost"].value < request.mode <= models.Mode["time"].value:
-        tes_times = estimate_times(
-            task_info=tes_task_info,
-            object_info=drs_object_info,
-            distances=tes_object_distances,
-        )
-    else: tes_times = {}
-    mode: float = float(request.mode)
-    # Rank by costs/times
-    # TODO: Should be method of class ServiceCombinations
-    ranked_services = rank_order.cost_time(
-        costs=tes_costs,
-        times=tes_times,
-        weight=request.mode,
-    )
-
-    # Randomize ranks
-    # TODO: Should be method of class ServiceCombinations
-    if mode == models.Mode["random"].value:
-        ranked_services = rank_order.randomize(
-            uris=tes_uris,
-            object_info=drs_object_info,
-        )
-    
-    # Catch other run modes
-    else:
-        logger.critical(
-            (
-                "Task cannot be computed: no ranking function defined for "
-                "mode: {mode}."
-            ).format(mode=mode)
-        )
-        raise ValueError
-
-    # Log output
-    log_yaml(
-        header="=== RANKED SERVICES ===",
-        level=logging.INFO,
-        logger=logger,
-        ranked_services=ranked_services,
-    )
-    
-
-    # MOVE TO OTHER MODULES & RE-FACTOR
-    #>>>>>>>>>>>>>>>>>>>>>>>>>>>
-    # tes_info_drs will have all the drs costs & info for each TES
-
-    cost_order = sorted(tes_costs.items(), key=lambda x: x[1]["total_costs"])
-    time_order = sorted(tes_task_info.items(), key=lambda x: x[1]["queue_time"]["duration"])
-
-    ranked_services = {uri: 0 for uri, val in cost_order}
-
-    # calculate the final rank on the basis of weight specified by mode
-    for i in range(0, len(cost_order)):
-        ranked_services[cost_order[i][0]] = ranked_services[cost_order[i][0]] + i*mode
-        ranked_services[time_order[i][0]] = ranked_services[time_order[i][0]] + i*(1-mode)
-
-    ranked_services = sorted(ranked_services.items(), key=lambda item: item[1])
-
-    # construct final return object
-    return_array_full = []
-    rank = 1
-    for i in ranked_services:
-        return_dict = {}
-
-        return_dict["access_uris"] = {}
-        return_dict["access_uris"]["tes_uri"] = i[0]
-        for drs_id in drs_object_info.keys():
-            return_dict["access_uris"][drs_id] = tes_info_drs[i[0]][drs_id][0]
-        return_dict["cost_estimate"] = {
-            "amount": tes_info_drs[i[0]]["total_costs"],
-            "currency": tes_info_drs[i[0]]["currency"]
-        }
-        return_dict["time_estimate"] = tes_task_info[i[0]]["queue_time"]
-        return_dict["rank"] = rank
-        return_array_full.extend([return_dict])
-        rank += 1
-
-    # Add warnings
-    response = {"warnings": [], "service_combinations": return_array_full}
-
+    # Return response
     return response
-    #<<<<<<<<<<<<<<<<<<<<<<<<<<<
+
+
+#    # Get valid TES-DRS combinations
+#    # TODO: Use model/class for return object
+#    # TODO: Add to constructor of Response
+#    valid_service_combos = get_valid_service_combinations(
+#        task_info=tes_task_info,
+#        object_info=drs_object_info,
+#    )
+#
+#    # Compute distances
+#    # TODO: Add to constructor of Response
+#    tes_object_distances = estimate_distances(
+#        combinations=valid_service_combos,
+#    )
+#
+#    # Filter service combinations
+#    # TODO: Add to constructor of Service Combinations
+#
+#    # Compute cost estimates
+#    # TODO: Should be method of class ServiceCombinations
+#    if models.Mode["cost"].value <= request.mode < models.Mode["time"].value:
+#        tes_costs = estimate_costs(
+#            task_info=tes_task_info,
+#            object_info=drs_object_info,
+#            distances=tes_object_distances,
+#        )
+#    else: tes_costs = {}
+#
+#    # Compute time estimates
+#    # TODO: Should be method of class ServiceCombinations
+#    if models.Mode["cost"].value < request.mode <= models.Mode["time"].value:
+#        tes_times = estimate_times(
+#            task_info=tes_task_info,
+#            object_info=drs_object_info,
+#            distances=tes_object_distances,
+#        )
+#    else: tes_times = {}
+#    mode: float = float(request.mode)
+#    # Rank by costs/times
+#    # TODO: Should be method of class ServiceCombinations
+#    ranked_services = rank_order.cost_time(
+#        costs=tes_costs,
+#        times=tes_times,
+#        weight=request.mode,
+#    )
+#
+#    # Randomize ranks
+#    # TODO: Should be method of class ServiceCombinations
+#    if mode == models.Mode["random"].value:
+#        ranked_services = rank_order.randomize(
+#            uris=tes_uris,
+#            object_info=drs_object_info,
+#        )
+#    
+#    # Catch other run modes
+#    else:
+#        logger.critical(
+#            (
+#                "Task cannot be computed: no ranking function defined for "
+#                "mode: {mode}."
+#            ).format(mode=mode)
+#        )
+#        raise ValueError
+#
+#    # Log output
+#    log_yaml(
+#        header="=== RANKED SERVICES ===",
+#        level=logging.INFO,
+#        logger=logger,
+#        ranked_services=ranked_services,
+#    )
+#    
+#
+#    # MOVE TO OTHER MODULES & RE-FACTOR
+#    #>>>>>>>>>>>>>>>>>>>>>>>>>>>
+#    # tes_info_drs will have all the drs costs & info for each TES
+#
+#    cost_order = sorted(tes_costs.items(), key=lambda x: x[1]["total_costs"])
+#    time_order = sorted(tes_task_info.items(), key=lambda x: x[1]["queue_time"]["duration"])
+#
+#    ranked_services = {uri: 0 for uri, val in cost_order}
+#
+#    # calculate the final rank on the basis of weight specified by mode
+#    for i in range(0, len(cost_order)):
+#        ranked_services[cost_order[i][0]] = ranked_services[cost_order[i][0]] + i*mode
+#        ranked_services[time_order[i][0]] = ranked_services[time_order[i][0]] + i*(1-mode)
+#
+#    ranked_services = sorted(ranked_services.items(), key=lambda item: item[1])
+#
+#    # construct final return object
+#    return_array_full = []
+#    rank = 1
+#    for i in ranked_services:
+#        return_dict = {}
+#
+#        return_dict["access_uris"] = {}
+#        return_dict["access_uris"]["tes_uri"] = i[0]
+#        for drs_id in drs_object_info.keys():
+#            return_dict["access_uris"][drs_id] = tes_info_drs[i[0]][drs_id][0]
+#        return_dict["cost_estimate"] = {
+#            "amount": tes_info_drs[i[0]]["total_costs"],
+#            "currency": tes_info_drs[i[0]]["currency"]
+#        }
+#        return_dict["time_estimate"] = tes_task_info[i[0]]["queue_time"]
+#        return_dict["rank"] = rank
+#        return_array_full.extend([return_dict])
+#        rank += 1
+#
+#    # Add warnings
+#    response = {"warnings": [], "service_combinations": return_array_full}
 
 
 # Executed when script is called from command line
 if __name__ == "__main__":
-    ranking = rank_services()
+    response = rank_services()
     log_yaml(
         header="=== OUTPUT ===",
         level=logging.INFO,
         logger=logger,
-        ranking=ranking,
+        ranking=response,
     )
+    # TODO: Do something with response
