--- TEStribute/config/server_config.yaml
@@ -1,27 +1,0 @@
-# General server settings
-server:
-    host: 0.0.0.0
-    port: 8080
-    debug: True
-
-# Security settings
-security:
-    authorization_required: False
-    jwt:
-        auth_header_key: Authorization
-        claim_identity: sub
-        claim_issuer: iss
-        claim_key_id: kid
-        decode_algorithms:
-          - RS256
-        idp_config_jwks: jwks_uri
-        idp_config_url_suffix: /.well-known/openid-configuration
-        idp_config_userinfo: userinfo_endpoint
-        jwt_prefix: Bearer
-        validation_methods:  # available methods: 'userinfo', 'public_key'
-          - userinfo
-          - public_key
-
-# OpenAPI specs
-openapi:
-    TEStribute: specs/schema.TEStribute.openapi.yaml
