+++ TEStribute/log/__init__.py
@@ -1,9 +1,12 @@
 """
-Configure logger.
+Logging configuration and convenience functions.
 """
 import logging
 import sys
-from typing import Union
+from typing import Optional
+
+import yaml
+
 
 def setup_logger(
     name: str,
@@ -35,3 +38,38 @@ def setup_logger(
         logger.addHandler(fl)
 
     return logger
+
+
+def log_yaml(
+    header: Optional[str] = None,
+    level: int = logging.DEBUG,
+    logger: logging.Logger = logging.getLogger(__name__),
+    **kwargs,
+) -> None:
+    """
+        Logs each of a number of keyword arguments with the indicated logging
+        level in YAML format. Dictionaries and iterable objects are logged
+        recursively.
+
+        :param header: If not `None`, the header is logged before any of the
+                keyword arguments are processed.
+        :param level: Logging level.
+        :param logger: The logger to be used.
+        :param separator: If not `None`, the separator is logged after every
+                keyword argument.
+
+        :return: None.
+    """
+    # Log header
+    if header is not None:
+        logger.log(level, header)
+
+    # Log value
+    if kwargs:
+        text = yaml.dump(
+            kwargs,
+            allow_unicode=True,
+            default_flow_style=False
+        ).splitlines()
+        for line in text:
+            logger.log(level, line)
