+++ TEStribute/models/__init__.py
@@ -3,9 +3,17 @@ Object models for representing nested, dependent data structures.
 """
 import enum
 import logging
-from typing import (Dict, Iterable, Union)
+from random import choice
+from typing import (Dict, Iterable, Mapping, Optional, Union)
 
-from TEStribute.errors import (ValidationError)
+from werkzeug.exceptions import Unauthorized
+
+from TEStribute.errors import (ResourceUnavailableError, ValidationError)
+from TEStribute.security.process_jwt import JWT
+from TEStribute.utils.service_calls import (
+    fetch_drs_objects_metadata,
+    fetch_tes_task_info,
+)
 
 logger = logging.getLogger("TEStribute")
 
@@ -233,6 +241,8 @@ class Request:
         drs_ids: Iterable[str] = [],
         drs_uris: Iterable[str] = [],
         mode: Union[float, int, Mode, str] = 0.5,
+        jwt: Optional[str] = None,
+        jwt_config: Mapping = {},
     ) -> None:
         """
         :param resource_requirements: Mapping of resources required for the
@@ -256,6 +266,16 @@ class Request:
                 in that case, the value determines the weight between cost (0)
                 and time (1) optimization.
         """
+        # Validate JWT
+        try:
+            JWT.config(**jwt_config)
+            jwt_obj = JWT(jwt=jwt)
+            jwt_obj.validate()
+        except Exception as e:
+            raise Unauthorized(e.args) from e
+
+        # Initiate instance variables
+        self.jwt = jwt_obj.jwt
         self.resource_requirements = resource_requirements
         self.tes_uris = tes_uris
         self.drs_ids = drs_ids
@@ -307,7 +327,6 @@ class Request:
 
     def sanitize_mode(
         self,
-        mode: Union[float, int, Mode, None, str] = None
     ) -> None:
         """
         Validates/sanitizes run mode.
@@ -347,7 +366,6 @@ class Request:
                 self.mode_float = self.mode
             else:
                 raise ValidationError(e)
-            
         
         # Raise ValidationError if mode is of any other type
         else:
@@ -366,10 +384,10 @@ class ServiceCombination:
         rank: int,
         time_estimate: Duration,
     ) -> None:
-        self.access_uris: access_uris
-        self.cost_estimate: cost_estimate
-        self.rank: rank
-        self.time_estimate: time_estimate
+        self.access_uris = access_uris
+        self.cost_estimate = cost_estimate
+        self.rank = rank
+        self.time_estimate = time_estimate
 
 
 class Response:
@@ -378,12 +396,224 @@ class Response:
     """
     def __init__(
         self,
-        service_combinations = Iterable[ServiceCombination],
-        warnings = Iterable[str]
-
+        request = Request,
+        timeout = float,
     ) -> None:
-        self.service_combinations: service_combinations
-        self.warnings: warnings
+        # TODO: Constructor needs to generate object conforming to valid
+        #       Response schema, but no values need to be computed yet
+
+        # Compile response object
+        self.service_combinations: Iterable[ServiceCombination]
+        self.warnings: Iterable[str] = []
+
+
+    # TODO: Implement instance methods to compute all required data for response
+    # - compute distances
+    # - filter service combinations
+    # - compute cost estimates
+    # - compute time estimates
+    # - calculate ranks
+    # - randomize ranks
+
+    @staticmethod
+    def get_valid_service_combinations(
+        resource_requirements: ResourceRequirements,
+        tes_uris: Iterable[str],
+        drs_ids: Iterable[str] = [],
+        drs_uris: Iterable[str] = [],
+        timeout: float = 3.0,
+        jtw: Optional[str] = None,
+    ) -> Dict:
+        """
+        Should take the similar input to the rak_services and return any random list of usable TES & DRS services
+
+        :task_info: Dict of the form
+        :object_info: Dict of form drs_uri's and drs_id's
+        :return: Dict of the form
+                {
+                    "tes_uri": [ # list of combinations for this TES
+                        {
+                            "a001": "object_uri",
+                            "a002": "object_uri",
+                            "output": "object_uri",
+                        }, {
+                            "a001": "object_uri",
+                            "a002": "object_uri",
+                            "output": "object_uri",
+                        }
+                    ],
+                    "tes_uri": [ # list of combinations for other TES
+                        {
+                            "a001": "object_uri",
+                            "a002": "object_uri",
+                            "output": "object_uri",
+                        }, {
+                            "a001": "object_uri",
+                            "a002": "object_uri",
+                            "output": "object_uri",
+                        }
+                    ]
+                }
+        """
+        # Get TES task info for resource requirements
+        try:
+            tes_task_info = fetch_tes_task_info(
+                tes_uris=tes_uris,
+                resource_requirements=resource_requirements,
+                timeout=timeout
+            )
+        except ResourceUnavailableError:
+            raise
+
+        # Get metadata for DRS input objects
+        try:
+            drs_object_info = fetch_drs_objects_metadata(
+                drs_uris=drs_uris,
+                drs_ids=drs_ids,
+                timeout=timeout
+            )
+        except ResourceUnavailableError:
+            raise
+
+        # Compile service combinations
+        return_dict = {}
+        for tes_uri in tes_task_info:
+            drs_dict = {}
+            for obj_id in drs_object_info:
+                drs_dict[obj_id] = choice(list(drs_object_info[obj_id].keys()))
+            return_dict[tes_uri] = drs_dict
+
+        return return_dict
+
+
+#def estimate_costs(
+#    task_info: Dict,
+#    object_info: Dict,
+#    distances: Dict,
+#) -> Dict:
+#    """
+#    """
+#    #############################
+#    tes_info_drs = {}
+#    for uri, info in task_info.items():
+#        # TODO : here the older values of each drs object's cost are overwitten
+#        #  fix the updation, though the normal dicts are NOT overwritten
+#        tes_info_drs.update({uri: sum_costs(
+#            total_tes_costs=task_info[uri]["costs_total"],
+#            data_transfer_rate=info["costs_data_transfer"],
+#            drs_objects_locations=object_info,
+#            tes_url=uri)
+#        })
+#    return tes_info_drs
+#    #############################
+#
+#
+#def transfer_costs(tes_url: str, rate: Dict, drs: Dict, size: float):
+#    """
+#    :param tes_url: string of the tes uri endpoint
+#    :param rate: Dict rate in format {"rate":,"currency:} rate in units per 1000 km
+#    :param drs: Dict of access methods
+#    :param size: Size of the drs object
+#    :return: Ordered dict of costs for the object according to cost
+#    """
+#    # considering more than one "access_methods" are provided
+#    for accessinfo in drs:
+#        distance = ip_distance(accessinfo["url"], tes_url)
+#        cost = round(size / 1000000000000, 7) * distance["distance"] * rate["amount"]
+#        accessinfo["cost"] = cost
+#        accessinfo["currency"] = rate["currency"]
+#
+#    return sorted(drs, key=lambda i: i["cost"])
+#
+#
+#def sum_costs(total_tes_costs: Dict, data_transfer_rate: Dict, drs_objects_locations: Dict, tes_url: str) -> Dict:
+#    """
+#    :param data_transfer_rate: cost from TES
+#    :param drs_objects_locations: Dict contaning the response from the fetch_tasks_info
+#    :param tes_url: url of the TES service w.r.t which drs is being computed
+#    :return:
+#    """
+#
+#    obj_size = defaultdict(dict)
+#    drs_info = defaultdict(dict)
+#    for obj_id, drs_uris in drs_objects_locations.items():
+#        # TO-DO:
+#        # consider the access method headers as well as checksums before cost estimation
+#        for drs_uri, access_info in drs_uris.items():
+#            # TO-DO:
+#            # look into checksums here & headers too for DRS services
+#            drs_info[obj_id][drs_uri] = [
+#                access_url["access_url"] for access_url in access_info["access_methods"]
+#            ]
+#            obj_size[obj_id][drs_uri] = access_info["size"]
+#
+#    return_info = defaultdict(dict)
+#    sum_drs = 0
+#    currency = None
+#    total_cost = None
+#    for drs_id, drs_info in drs_info.items():
+#        for drs_uri, object_info in drs_info.items():
+#            # only the one with the lowest cost is kept
+#            return_info[drs_id][drs_uri] = transfer_costs(
+#                tes_url, data_transfer_rate, object_info, obj_size[drs_id][drs_uri]
+#            )[0]
+#
+#        # return only cheapest drs
+#        return_info[drs_id] = sorted(
+#            return_info[drs_id].items(), key=lambda x: x[1]["cost"]
+#        )[0]
+#
+#        sum_drs = return_info[drs_id][1]["cost"] + sum_drs
+#        currency = return_info[drs_id][1]["currency"]
+#
+#        if currency == total_tes_costs["currency"]:
+#            total_cost = total_tes_costs["amount"] + sum_drs
+#        else:
+#            raise ValueError
+#
+#    if currency is not None and total_cost is not None:
+#
+#        # to fix overwriting of elemets in the loop it is called from
+#        # return_info = dict(return_info)
+#        return_info.update({
+#            "drs_costs": {"amount": sum_drs, "currency": currency},
+#            "total_costs": total_cost,
+#            "currency": currency
+#        })
+#
+#        return return_info
+#
+#    else:
+#        raise ValueError
+#
+#
+#def estimate_times(
+#    task_info: Dict,
+#    object_info: Dict,
+#    distances: Dict,
+#) -> Dict:
+#    """
+#    """
+#    return {}
+#
+#
+#def randomize(
+#    uris: Iterable,
+#    object_info: Dict,
+#) -> Dict:
+#    """
+#    """
+#    return {}
+#
+#
+#def cost_time(
+#    costs: Dict,
+#    times: Dict,
+#    weight: float,
+#) -> Dict:
+#    """
+#    """
+#    return {}
 
 
 class TaskInfo:
