--- TEStribute/rank_order.py
@@ -1,22 +1,0 @@
-"""
-"""
-from typing import (Dict, Iterable)
-
-
-def randomize(
-    uris: Iterable,
-    object_info: Dict,
-) -> Dict:
-    """
-    """
-    return {}
-
-
-def cost_time(
-    costs: Dict,
-    times: Dict,
-    weight: float,
-) -> Dict:
-    """
-    """
-    return {}
