--- TEStribute/times.py
@@ -1,14 +1,0 @@
-"""
-Functions to compute times associated with tasks.
-"""
-from typing import Dict
-
-
-def estimate_times(
-    task_info: Dict,
-    object_info: Dict,
-    distances: Dict,
-) -> Dict:
-    """
-    """
-    return {}
